fixed multer issue

+++ routes/admission.js
@@ -3,9 +3,10 @@ const router = express.Router();
 const multer = require("multer")
 const chalk = require('chalk');
 const { body, validationResult } = require("express-validator")
-const { admission } = require("../models")
+const { admission, guardian } = require("../models")
 const models = require("../models");
 var path = require("path");
+const fs = require("fs");
 
 const validate = [
   body("firstName")
@@ -23,7 +24,7 @@ const validate = [
   .isLength({ min: 1, max: 50 })
   .withMessage("last name should be atleast 1 char and atmost 20 chars long"),
   body("academicYear")
-  .isDate()
+  .isString()
   .withMessage("academic year should be a date"),
   body("registerNo")
   .isString()
@@ -31,35 +32,45 @@ const validate = [
   .trim()
   .isLength({ min: 1, max: 255 }),
   body("admissionDate")
-  .isDate()
+  .isString()
   .withMessage("admission date should be a date")
 ]
 
 const storage = multer.diskStorage({
   destination: './public/uploads/studentImages',
-  filename: function(req, file, cb){
-    cb(null,file.fieldname + '-' + Date.now() + path.extname(file.originalname));
+  filename: function (req, file, cb) {
+    const datetimestamp = Date.now()
+    cb(null, file.fieldname + '-' + datetimestamp + '.' + file.originalname.split('.')[file.originalname.split('.').length -1])
   }
 });
 
 
 const upload = multer({
   storage: storage,
-  limits:{fileSize: "1000000"},
- /* fileFilter: function(req, file, cb){
-  const fileTypes = /jpeg|jpg|png/
-    const mimeType = fileTypes.test(file.mimetype)
-    const extname = fileTypes.test(path.extname(file.originalname))
-
-    if(mimeType && extname) {
-        return cb(null, true)
-    }
-    cb('warning!!! only images can be uploaded')
-  }*/
-}).single('photo');
+  limits:{fileSize: 1000000},
+  fileFilter: function(req, file, cb){
+    checkFileType(file, cb)
+  }
+}).single('photo')
 
 
-//route to list vendors
+// Check File Type
+function checkFileType(file, cb){
+  // Allowed ext
+  const filetypes = /jpeg|jpg|png/;
+  // Check ext
+  const extname = filetypes.test(path.extname(file.originalname).toLowerCase());
+  // Check mime
+  const mimetype = filetypes.test(file.mimetype);
+
+  if(mimetype && extname){
+    return cb(null,true);
+  } else {
+    cb(console.log("only images are allowed"));
+  }
+}
+
+
 router.get("/list", async function(req, res){
 
   try {
@@ -89,7 +100,7 @@ router.get("/list", async function(req, res){
 });
 
 
-router.post("/add",upload, async function(req, res) {
+router.post("/add", upload, validate, async function(req, res) {
 
   try {
     console.log(req.body)
@@ -113,10 +124,64 @@ router.post("/add",upload, async function(req, res) {
         })
       })
     }else {
+      
+      if(req.body.guardianId){
+        var guardianId = guardianId
+      }else {
+        const {
+          name,
+          relation,
+          fatherName,
+          motherName,
+          occupation,
+          income,
+          education,
+          city,
+          state,
+          mobile,
+          email,
+          address,
+        } = await req.body
+
+      transaction = await models.sequelize.transaction()
+
+      var { id } = await guardian.create({
+        name,
+        relation,
+        fatherName,
+        motherName,
+        occupation,
+        income,
+        education,
+        city,
+        state,
+        mobile,
+        email,
+        address
+      },
+      {
+        transaction
+      }).catch(err => {
+        message = "add fail"
+        status = 500
+        success = false
+        console.log(chalk.red.bold(err.message))
+      })
+
+      if(!success){
+        transaction.rollback()
+      }else {
+        transaction.commit()
+      }
+    }
+
+    if(success) {
+      guardianId = id
 
       const {
-        section,
-        std,
+        classRoomId,
+        sectionId,
+        guardianId,
         academicYear,
         registerNo,
         admissionDate,
@@ -136,19 +201,16 @@ router.post("/add",upload, async function(req, res) {
         state,
         presentAddress,
         permanentAddress,
-        transportRoute,
-        vehicleNo,
-        hostelName,
-        roomNo,
+        studentCategoryId,
         previousSchoolName,
         qualification,
         remarks
-      } = await req.body;
-      const { photo } = req.file.path
-      transaction = await models.sequelize.transaction()
+      } = await req.body
+
       await admission.create({
-        section,
-        std,
+        sectionId,
+        classRoomId,
+        guardianId,
         academicYear,
         registerNo,
         admissionDate,
@@ -168,14 +230,15 @@ router.post("/add",upload, async function(req, res) {
         state,
         presentAddress,
         permanentAddress,
-        photo,
         transportRoute,
         vehicleNo,
         hostelName,
         roomNo,
+        studentCategoryId,
         previousSchoolName,
         qualification,
-        remarks
+        remarks,
+        photo: req.file.path,
       },
       {
         transaction
@@ -189,9 +252,9 @@ router.post("/add",upload, async function(req, res) {
         await transaction.rollback()
       }else {
         await transaction.commit()
-
       }
     }
+  }
   }catch (err) {
     success = false
     message = "add fail"
@@ -206,7 +269,6 @@ router.post("/add",upload, async function(req, res) {
 });
 
 
-//route to delete a vendor
 router.delete("/delete/:id", async function(req, res){
 
   try {
@@ -214,22 +276,43 @@ router.delete("/delete/:id", async function(req, res){
     var message = "delete success"
     var status = 200
     const id = req.params.id;
-    await admission.destroy({ 
-      where: {
-        id
-      }
-    }).catch(err => {
-      success = false
-      message = "delete fail"
-      status = 500 
+    const data = await admission.findByPk(id).catch(err => {
       console.log(err.message)
     })
+    console.log("data dfsladfasfd",data)
+    if(data){
+      console.log("im in here")
+      await admission.destroy({ 
+        where: {
+          id
+        }
+      }).catch(err => {
+        success = false
+        message = "delete fail"
+        status = 500 
+        console.log(err.message)
+      })
+      if(success){
+        let resultHandler = function (err) {
+          if (err) {
+            console.log("file delete failed", err);
+          } else {
+            console.log("file deleted");
+          }
+        }
+      fs.unlink(data.photo, resultHandler);
+      }
+    }else{
+      success = false
+      message = "delete fail"
+      status = 400
+    }
 
   } catch (err) {
     success = false
     message = "delete fail"
     status = 400
-    console.log(err);
+    console.log(err.message)
   };
 
   res.status(status).json({
@@ -294,7 +377,7 @@ router.patch("/edit/:id", validate, async function(req, res){
         previousSchoolName,
         qualification,
         remarks
-      } = await req.body;
+      } = await req.body
       transaction = await models.sequelize.transaction()
       await admission.update({
         section,
@@ -325,7 +408,8 @@ router.patch("/edit/:id", validate, async function(req, res){
         roomNo,
         previousSchoolName,
         qualification,
-        remarks
+        remarks,
+        photo: req.file.path
       },
       {
         where: { id }
@@ -351,14 +435,13 @@ router.patch("/edit/:id", validate, async function(req, res){
     success = false
     message = "edit fail"
     status = 400
-    console.log(err);
+    console.log(err)
   };
   res.status(status).json({
     success,
     message,
     validationError
   })
-});
+})
 
-module.exports = router;
-
+module.exports = router
