+++ routes/deposit.js
@@ -4,6 +4,9 @@ const chalk = require('chalk');
 const { body, validationResult } = require("express-validator")
 const { deposit, voucherHead, account } = require("../models");
 const models = require("../models");
+const fs = require("fs")
+const multer = require('multer')
+const path = require("path")
 
 const validate = [
   body("ref")
@@ -13,7 +16,7 @@ const validate = [
   .isLength({ min: 1, max: 255 })
   .withMessage("ref should be atleast 1 char and atmost 20 chars long"),
   body("amount")
-  .isNumeric()
+  .isString()
   .withMessage("amount should be a number"),
   body("date")
   .isString()
@@ -21,8 +24,42 @@ const validate = [
   body("payVia")
   .isString()
   .withMessage("payment method is required")
-  .isIn(["Cash", "Card", "Cheque", "Bank Transfer"])
 ]
+
+const storage = multer.diskStorage({
+  destination: './public/uploads/depositImages',
+  filename: function (req, file, cb) {
+    const datetimestamp = Date.now()
+    console.log("inside storge")
+    cb(null, file.fieldname + '-' + datetimestamp + '.' + file.originalname.split('.')[file.originalname.split('.').length -1])
+  }
+});
+
+
+const upload = multer({
+  storage: storage,
+  limits:{fileSize: 1000000},
+  fileFilter: function(req, file, cb){
+    checkFileType(file, cb)
+  }
+}).single('attachment')
+
+
+// Check File Type
+function checkFileType(file, cb){
+  // Allowed ext
+  const filetypes = /jpeg|jpg|png/
+  // Check ext
+  const extname = filetypes.test(path.extname(file.originalname).toLowerCase())
+  // Check mime
+  const mimetype = filetypes.test(file.mimetype)
+
+  if(mimetype && extname){
+    return cb(null,true)
+  } else {
+    cb('Error: Images Only!')
+  }
+}
 
 //route to list vendors
 router.get("/list", async function(req, res){
@@ -66,7 +103,7 @@ router.get("/list", async function(req, res){
 
 
 //route to add a vendor
-router.post("/add", validate, async function(req, res) {
+router.post("/add", upload, validate, async function(req, res) {
 
   try {
 
@@ -107,7 +144,8 @@ router.post("/add", validate, async function(req, res) {
         amount,
         date,
         payVia,
-        description
+        description,
+        attachment: req.file.path
       },
       {
         transaction
@@ -121,7 +159,6 @@ router.post("/add", validate, async function(req, res) {
         await transaction.rollback()
       }else {
         await transaction.commit()
-
       }
     }
   }catch (err) {
@@ -146,23 +183,41 @@ router.delete("/delete/:id", async function(req, res){
     var message = "delete success"
     var status = 200
     const id = req.params.id;
-    await deposit.destroy({ 
-      where: {
-        id
-      }
-    }).catch(err => {
-      success = false
-      message = "delete fail"
-      status = 500 
+    const data = await deposit.findByPk(id).catch(err => {
       console.log(err.message)
     })
-
+    if(data){
+      await deposit.destroy({ 
+        where: {
+          id
+        }
+      }).catch(err => {
+        success = false
+        message = "delete fail"
+        status = 500 
+        console.log(err.message)
+      })
+      if(success){
+        let resultHandler = function (err) {
+          if (err) {
+            console.log("file delete failed", err);
+          }else {
+            console.log("file deleted");
+          }
+        }
+      fs.unlink(data.attachment, resultHandler);
+      }
+    }else{
+      success = false
+      message = "delete fail"
+      status = 400
+    }
   } catch (err) {
     success = false
     message = "delete fail"
     status = 400
-    console.log(err);
-  };
+    console.log(err.message)
+  }
 
   res.status(status).json({
     success,
@@ -172,7 +227,7 @@ router.delete("/delete/:id", async function(req, res){
 
 
 //route to update a vendor
-router.patch("/edit/:id", validate, async function(req, res){
+router.patch("/edit/:id", upload, validate, async function(req, res){
 
   try {
 
@@ -204,7 +259,7 @@ router.patch("/edit/:id", validate, async function(req, res){
         date,
         payVia,
         description
-      } = await req.body;
+      } = await req.body
       transaction = await models.sequelize.transaction()
       await deposit.update({ 
         accountId,
@@ -213,7 +268,8 @@ router.patch("/edit/:id", validate, async function(req, res){
         amount,
         date,
         payVia,
-        description
+        description,
+        attachment: req.file.path
       },
       {
         where: { id }
@@ -224,20 +280,20 @@ router.patch("/edit/:id", validate, async function(req, res){
         message = "edit fail"
         status = 500
         success = false
-        console.log(chalk.red.bold("im from first inner catch",err.message))
+        console.log(chalk.red.bold(err.message))
       })
       if(!success) {
         await transaction.rollback()
       }else {
         await transaction.commit()
-      }    
+      }
     }
   }catch (err) {
     success = false
     message = "edit fail"
     status = 400
-    console.log(err);
-  };
+    console.log(err)
+  }
   res.status(status).json({
     success,
     message,
